+++ src/main/java/com/nexmo/client/voice/endpoints/ModifyCallMethod.java
@@ -31,6 +31,7 @@ import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
 import org.apache.http.entity.StringEntity;
+import org.apache.http.impl.client.BasicResponseHandler;
 import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
@@ -64,8 +65,7 @@ public class ModifyCallMethod extends AbstractMethod<CallModifier, ModifyCallRes
 
     @Override
     public ModifyCallResponse parseResponse(HttpResponse response) throws IOException {
-        String json = EntityUtils.toString(response.getEntity());
-        LOG.info("Received: " + json);
+        String json = new BasicResponseHandler().handleResponse(response);
         return ModifyCallResponse.fromJson(json);
     }